refactor: MyPage 컴포넌트 Common 컴포넌트로 리팩토링
